+++ lib/output/sql_extra.go
@@ -1,0 +1,9 @@
+// +build !wasm
+
+package output
+
+// Import extra drivers that aren't supported by WASM builds.
+import (
+	// SQL Drivers
+	_ "github.com/lib/pq"
+)